changed longint by size_t
git-svn-id: file:///home/roberto/Git/svn@7621 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 45e4e020fd9a3952bb1e18239b21dc019741f468

+++ trunk/xmipp/libraries/reconstruction/xray_import.cpp
@@ -271,7 +271,7 @@ void runThread(ThreadArgument &thArg)
     MetaData localMD;
     Image<double> Iaux;
     FileName fnImg;
-    long long int first = -1, last = -1;
+    size_t first = -1, last = -1;
     MultidimArray<char> mask;
 
     while (ptrProg->td->getTasks(first, last))
